(CDAP-5800) Combine multiple user preferences test classes into one
- Reduce the overhead in start/stop app-fabric multiple times
